fix(NP-1127): add 'imagePlain' class to image plain template
After some sleuthing I was able to find that there was already a CSS
rule written to remove the padding from imagePlain events. Going back
to as far as v4.8.76 I was not able to find usages of the class so
I assume it was written and never actually applied to the template in
question.