updates(audio-mode): enable audio-only mode in layouts.
hide_video is a boolean value added to scene assets which is set from
producer edit layout tab.

The video is hidden (itt-magnet) using an ng-hide to keep it around
but remove it from view. The itt-magnetized is removed from the DOM with
an ng-if in order to allow the surrounding content in the layout to
slide up.