fix(NP-1047): remove deprecated noExternalLink event property
outgoing link display is handled on EditController. noExternalLink was
initialized as false, then set to 'true' in the case that the template
contained the string '-embed' in its templateUrl field.

Now that we handle outgoing links based upon the results of the
x-frame-options proxy, we can safely remove all instances of this field